requirements.txt
Flask==2.2.3
redis==4.5.3
psycopg2-binary==2.9.6
